% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplots-internal.R
\name{xplots-internal}
\alias{xplots-internal}
\alias{plotUG}
\alias{plotPG}
\title{Internal xplots functions}
\usage{
plotUG(marks, plotTitle, zeros, altAssRows, alt)

plotPG(marks, plotTitle, zeros, altAssRows, alt)
}
\description{
Internal xplots functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
