% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplots-package.R
\docType{package}
\name{xplots-package}
\alias{xplots}
\alias{xplots-package}
\title{xplots: functions to produce plots of marks for an exam board.}
\description{
Example code for producing
\itemize{
\item pairs plots of overall marks from different modules, and
\item pairs plots of component marks for given modules.
}
}
\details{
The main functions are
\itemize{
\item \code{\link{components}}: plots to compare components marks for a given module.
\item \code{modules}: plots to compare module marks within groups of modules.
(\strong{Not available yet!})
}
}
\author{
\strong{Maintainer}: Paul Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

}
