% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\title{By-module scatter plots of component marks}
\usage{
components(
  input = "marks",
  output = "plots",
  skip = 9,
  filename = "components.pdf",
  zeros = FALSE,
  alt = c("red", "black", "exclude")
)
}
\arguments{
\item{input}{The directory containing the module mark Excel files.}

\item{output}{The directory in which the PDF file of plots is to be written.
If this does not exist then it will be created.}

\item{skip}{An integer scalar. The number of rows to skip before reading
data from the source marks file, that is, the number of rows before the
row that contains the column names.}

\item{filename}{The name of the output PDF file.}

\item{zeros}{A logical scalar. Should marks of zero be included in the
plots?}

\item{alt}{A character scalar that determines how marks for alternative
assessments are plotted
\itemize{
\item \code{"red"}: to highlight these marks against the black used for other
marks,
\item \code{"black"}: to plot in black,
\item \code{"exclude"}: to exclude from the plots.
}}
}
\value{
A named list containing the component marks for each module is
returned invisibly. The names are the module codes.
}
\description{
Create a PDF file containing, for each module, a pairs plot that compares
module component marks.
}
\details{
A single PDF file is produced. Each page contains a matrix of
scatterplots, produced using \code{\link[graphics:pairs]{graphics::pairs}}, in which the component
marks for a module are plotted against each other. There is a page for
each spreadsheet of module marks in the \code{input} directory. The title of a
plot is the module code followed by the module delivery code, e.g.
HPSC0003 A4U.

The relevant pass mark and the first/distinction mark of 70\% are indicated
by dotted lines on the plots. For postgraduate modules (A7P) the area
where at least one mark is below the lower end of the condonable range
(40\%) is shaded grey and any component mark below 40\% is plotted using
a special symbol, a circle with a cross on it.

If an error is thrown when trying to produce a plot for a module then a
warning message is printed and this module is skipped. If a module has
only one assessment component then this module also skipped.

\strong{Notes}
\itemize{
\item The files \code{HPSC0003.xls} etc, which I know have been downloaded direct
from Portico and advertised as Excel files, seem \strong{not} to be Excel
files, but rather formatted as webpages (\code{.html} or the equivalent). This
meant that I could not read their contents into R. To solve this, I opened
them in Excel and saved them as Excel Workbook (*.xlsx). These module mark
Excel files, e.g., \code{HPSC0003.xlsx}, should be placed in the directory
\code{marks}.
\item It is assumed that a given \code{.xlsx} file does not mix students of Levels
4-6 with students of Level 7. The level is inferred by searching for the
Module Delivery entry in the \code{.xlsx} file. If there are modules taken by
students of Level 6 and Level 7 then their marks should be in separate
files, for example, \code{HPSC0003.xlsx} and \code{HPSC0003level7.xlsx}.
\item The example file \code{HPSC0007.xlsx} contained marks for an alternative
assessments, separated from the main set of marks. The code looks for this
and merges these marks with the others, under the assumption that there is
the same number of assessments as in the main set of marks and that they
are in a comparable order. The argument \code{alt} determines how/whether these
marks are plotted.
}
}
\examples{
\dontrun{
# The following code will produce a PDF file containing a page of plots for
# each spreadsheet of module marks in the input directory
components()
}

}
